# openapi.yaml
openapi: 3.0.3
info:
  title: track-web API
  description: Self-hosted time tracking, watch tracking, trips, and social API.
  version: 1.0.0
servers:
  - url: https://time.branam.us
    description: Time tracking app
  - url: https://watch.branam.us
    description: Watch tracking app
  - url: https://trips.branam.us
    description: Trips app
  - url: https://proto.branam.us
    description: Proto app
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
      required: [ok]
...
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        appId:
          type: string
        description:
          type: string
        tags:
          type: string
          description: Comma-separated bare tag words (no prefix)
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, appId, description, tags, startedAt]
    Trip:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        destination:
          type: string
          nullable: true
        departureNotes:
          type: string
          nullable: true
        returnNotes:
          type: string
          nullable: true
        nights:
          type: integer
          nullable: true
        fullDays:
          type: integer
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        infoMarkdown:
          type: string
          nullable: true
        isCurrent:
          type: boolean
      required: [id, userId, name]
...
    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        runtimeMinutes:
          type: integer
          nullable: true
        releaseYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        streaming:
          type: string
          nullable: true
        tmdbId:
          type: integer
          nullable: true
        addedByUserId:
          type: integer
      required: [id, title, addedByUserId]
    TvSeries:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        streaming:
          type: string
          nullable: true
        episodeRuntimeMinutes:
          type: integer
          nullable: true
        seasonCount:
          type: integer
          nullable: true
        releaseYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        tmdbId:
          type: integer
          nullable: true
        addedByUserId:
          type: integer
      required: [id, title, addedByUserId]
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
paths:
...
  /api/llm-context.md:
    get:
      summary: Get LLM agent context guide
      description: Returns a Markdown document describing the app's purpose, auth model, feature areas, and key conventions. Intended for LLM agents to load as context before using the API.
      operationId: getLlmContext
      security: []
      responses:
        '200':
          description: Markdown context document
          content:
            text/markdown:
              schema:
                type: string
        '404':
          description: Context file not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/openapi.json:
    get:
      summary: Get OpenAPI spec
      description: Returns the parsed OpenAPI spec as JSON. No authentication required.
      operationId: getOpenApiSpec
      security: []
      responses:
        '200':
          description: OpenAPI document as JSON
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Spec file not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Auth ────────────────────────────────────────────────────────────────────
  /api/auth/login:
    post:
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
              required: [email, password]
      responses:
        '200':
          description: Login successful; sets session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many failed attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/time/entries:
    get:
      summary: List entries for a day
      operationId: listEntries
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          description: Date in YYYY-MM-DD format (defaults to today in ET using 4 AM boundary)
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        '200':
          description: Entries for the day
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimeEntry'
                required: [entries]
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid date format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Start a new time entry
      operationId: createEntry
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  minLength: 1
                  maxLength: 500
                startedAt:
                  type: string
                  format: date-time
              required: [description, startedAt]
      responses:
        '201':
          description: Entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  entry:
                    $ref: '#/components/schemas/TimeEntry'
                required: [entry]
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: A task is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Start time conflicts with previous entry
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  previousEndedAt:
                    type: string
                    format: date-time
...
  /api/trips:
    get:
      summary: List all trips
      operationId: listTrips
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of trips
          content:
            application/json:
              schema:
                type: object
                properties:
                  trips:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trip'
                required: [trips]
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a trip
      operationId: createTrip
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                destination:
                  type: string
                  maxLength: 200
                  nullable: true
                departureNotes:
                  type: string
                  maxLength: 2000
                  nullable: true
                returnNotes:
                  type: string
                  maxLength: 2000
                  nullable: true
                nights:
                  type: integer
                  minimum: 0
                  nullable: true
                fullDays:
                  type: integer
                  minimum: 0
                  nullable: true
                startDate:
                  type: string
                  format: date
                  nullable: true
                endDate:
                  type: string
                  format: date
                  nullable: true
                infoMarkdown:
                  type: string
                  nullable: true
              required: [name]
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                type: object
                properties:
                  trip:
                    $ref: '#/components/schemas/Trip'
                required: [trip]
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/watch/tags:
    get:
      summary: List watch tags
      operationId: listWatchTags
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create a watch tag
      operationId: createWatchTag
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
              required: [name]
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '409':
          description: Tag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Watch: Movies ─────────────────────────────────────────────────────────────
...
  /api/watch/movies:
    get:
      summary: List movies
      operationId: listMovies
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Create a movie
      operationId: createMovie
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                runtimeMinutes:
                  type: integer
                  nullable: true
                releaseYear:
                  type: integer
                  nullable: true
                description:
                  type: string
                  nullable: true
                streaming:
                  type: string
                  nullable: true
                tagIds:
                  type: array
                  items:
                    type: integer
              required: [title]
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
...
  /api/watch/tv:
    get:
      summary: List TV series
      operationId: listTvSeries
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of TV series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TvSeries'
    post:
      summary: Create a TV series
      operationId: createTvSeries
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                streaming:
                  type: string
                  nullable: true
                episodeRuntimeMinutes:
                  type: integer
                  nullable: true
                seasonCount:
                  type: integer
                  nullable: true
                releaseYear:
                  type: integer
                  nullable: true
                description:
                  type: string
                  nullable: true
                tagIds:
                  type: array
                  items:
                    type: integer
              required: [title]
      responses:
        '201':
          description: TV series created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TvSeries'
...
  /api/watch/events:
    get:
      summary: List watch events
      operationId: listWatchEvents
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [active, completed-recent]
      responses:
        '200':
          description: List of watch events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      summary: Create a watch event
      operationId: createWatchEvent
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                scheduledDate:
                  type: string
                  minLength: 1
                invitees:
                  type: array
                  items:
                    oneOf:
                      - type: object
                        properties:
                          type:
                            type: string
                            enum: [user]
                          userId:
                            type: integer
                        required: [type, userId]
                      - type: object
                        properties:
                          type:
                            type: string
                            enum: [group]
                          groupId:
                            type: integer
                        required: [type, groupId]
                  default: []
              required: [title, scheduledDate]
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
...
  /api/watch/ratings:
    get:
      summary: Get all personal ratings (movies + TV merged)
      operationId: listRatings
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Ratings sorted by score descending
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # ── Watch: External Search & Import ──────────────────────────────────────────
...
  /api/deploy:
    post:
      summary: GitHub webhook deploy trigger
      description: Verified via HMAC-SHA256 using x-hub-signature-256 header. Only triggers on refs/heads/main.
      operationId: deployWebhook
      security: []
      parameters:
        - name: x-hub-signature-256
          in: header
          required: true
          schema:
            type: string
            description: "HMAC-SHA256 of the request body, format: sha256=<hex>"
      responses:
        '200':
          description: Push was not to main branch; no deploy triggered
        '202':
          description: Deploy triggered
        '403':
          description: Invalid signature
        '503':
          description: DEPLOY_SECRET not configured
  /api/admin/deploy:
    post:
      summary: Trigger a server deploy (admin only)
      description: Admin app deploy. Requires a session for user 1. Replaces the former /api/deploy/trigger.
      operationId: adminDeploy
      security:
        - cookieAuth: []
      responses:
        '202':
          description: Deploy triggered
        '401':
          description: Not authenticated (session required)
        '403':
          description: Not an admin (userId must be 1)
...
  /api/scores:
    post:
      summary: Submit a game score
      description: Records the authenticated user's score for a completed game session.
      operationId: submitScore
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gameSlug, mode, level, score]
              properties:
                gameSlug:
                  type: string
                  example: ball-merge
                mode:
                  type: string
                  example: classic
                level:
                  type: string
                  example: box
                score:
                  type: integer
                  minimum: 1
                  description: Must be a positive integer; score of 0 is rejected with 422
      responses:
        '201':
          description: Score recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  userId: { type: integer }
                  gameSlug: { type: string }
                  mode: { type: string }
                  level: { type: string }
                  score: { type: integer }
                  achievedAt: { type: string, format: date-time }
        '401': { description: Not authenticated }
        '422': { description: Invalid payload }
...